query: Allow multiple filters to be specified
To make the query API more flexible, while retaining simplicity for simple
queries, allow multiple filters to be specified in a query. Objects returned in
the query response must match all criteria in at least one of the specified
filters; i.e., within a given filter the criteria are a conjunction (AND), while
the multiple filters are a disjunction (OR). This is a breaking change, hence
the version bump to v1alpha2.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>